requirements.txt
torch~=2.0.0
pygame~=2.5.0
numpy~=1.22.4
pandas~=1.5.3
scikit-learn~=1.2.2
tensorboardX~=2.5.1
transformers~=4.27.1
matplotlib~=3.5.2
opencv-python~=4.6.0.66
scikit-image~=0.18.1
tqdm~=4.64.0
einops~=0.6.1
torchvision~=0.15.0
Pillow~=9.0.1
seaborn~=0.12.2
timm~=0.9.2
scipy~=1.7.3
fastai~=1.0.61
openpyxl~=3.0.10
aiohttp~=3.8.4
requests~=2.28.1
nltk~=3.7
rouge~=1.0.1
evaluate~=0.4.0
EasyNMT~=2.0.2
wfdb~=4.1.1